streamer/requirements.txt
kafka-python==2.0.2; python_version < '3.12'
kafka-python @ git+https://github.com/dpkp/kafka-python.git ; python_version >= '3.12'
requests==2.32.3
websocket-client==1.8.0
python-dotenv==1.0.1
colorama
pytest